feat: create LMChatMediaManager class
